requirements.txt
openai>=0.27.0
pinecone-client>=2.0.0
sentence-transformers>=2.2.2
beautifulsoup4>=4.11.1
lxml>=4.9.2
html2text>=2020.1.16
pandas>=1.5.3
requests>=2.28.0
